# Conflicts:
#	armor-view/src/components/Notice.vue
#	armor-view/src/components/Stats2.vue

+++ armor-api/src/main/java/com/zhixinhuixue/armor/model/pojo/OnlineQuestion.java
@@ -29,7 +29,7 @@ public class OnlineQuestion {
     /**
      * 工作时长
      */
-    private Integer workHour;
+    private Float workHour;
 
     /**
      * 开始时间
@@ -103,11 +103,11 @@ public class OnlineQuestion {
         this.projectId = projectId;
     }
 
-    public Integer getWorkHour() {
+    public Float getWorkHour() {
         return workHour;
     }
 
-    public void setWorkHour(Integer workHour) {
+    public void setWorkHour(Float workHour) {
         this.workHour = workHour;
     }
